pyproject.toml
[project]
name = "lkr"
version = "0.1.0"
description = "Git-native, LLM-ready knowledge repository manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "rich>=13.0",
    "pydantic>=2.5",
]

[project.scripts]
lkr = "lkr.cli:cli"
k = "lkr.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lkr"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=7.0",
    "pytest-xdist>=3.8",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
